+++ src/tree_view/items/space_item.ts
@@ -1,13 +1,14 @@
 import path = require('path');
 import { TreeItem, TreeItemCollapsibleState } from 'vscode';
+import { Space } from '../../types';
 
 export class SpaceItem extends TreeItem {
     constructor(
-        public id: string,
-        public readonly label: string,
+        public space: Space,
         public readonly collapsibleState: TreeItemCollapsibleState
     ) {
-        super(label, collapsibleState);
+        super(space.name, collapsibleState);
+        this.id = space.id;
         this.iconPath = {
             light: path.join(__filename, '..', '..', '..', '..', 'resources', 'spaceItem', 'layers.png'),
             dark: path.join(__filename, '..', '..', '..', '..', 'resources', 'spaceItem', 'layers_white.png')